Change type of id from Long to BigInteger - pass test
- https://docs.spring.io/spring-data/mongodb/docs/current/reference/html/#mongo-template.id-handling

+++ biz/seller/src/main/java/io/homo_efficio/msa/simple_mall/biz/seller/domain/repository/SellerRepository.java
@@ -3,9 +3,11 @@ package io.homo_efficio.msa.simple_mall.biz.seller.domain.repository;
 import io.homo_efficio.msa.simple_mall.biz.seller.domain.model.Seller;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 
+import java.math.BigInteger;
+
 /**
  * @author homo.efficio@gmail.com
  * created on 2020-06-29
  */
-public interface SellerRepository extends ReactiveMongoRepository<Seller, Long> {
+public interface SellerRepository extends ReactiveMongoRepository<Seller, BigInteger> {
 }
